requirements.txt
asgiref==3.2.10
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==3.0.3
channels_redis==4.2.1
constantly==23.10.4
cryptography==45.0.4
daphne==3.0.2
Django==3.1
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
pytz==2025.2
redis==6.2.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.14.0
zope.interface==7.2
